+++ src/main/kotlin/net/thoughtmachine/please/plugin/runconfiguration/PleaseLineMarkerProvider.kt
@@ -68,43 +68,49 @@ object PleaseLineMarkerProvider : RunLineMarkerContributor() {
     private fun filterActions(project: Project, target: BuildTarget): List<AnAction> {
         val label = target.toString()
         val actions = mutableListOf(PleaseAction(project, label, "build",
-            PleaseBuildExecutor, newBuildConfig(project, label)))
+            PleaseBuildExecutor, newBuildConfig(project, target.pkg.pleaseRoot, label)))
 
-        if (target.test) {
+        val test = target.info?.test ?: true
+        val binary = target.info?.binary ?: true
+
+        if (test) {
             actions.addAll(listOf(
                 PleaseAction(project, label, "test",
-                    DefaultRunExecutor.getRunExecutorInstance(), newTestConfig(project, label)),
+                    DefaultRunExecutor.getRunExecutorInstance(), newTestConfig(project, target.pkg.pleaseRoot, label)),
                 PleaseAction(project, label, "test",
-                    DefaultDebugExecutor.getDebugExecutorInstance(), newTestConfig(project, label)),
+                    DefaultDebugExecutor.getDebugExecutorInstance(), newTestConfig(project, target.pkg.pleaseRoot, label)),
             ))
-        } else if (target.binary) {
+        }
+        // Include these options if we don't have any information about the target. This seems more useful. If it's not
+        // a test or binary, Please will give a decent error message anyway.
+        if (binary && !test || target.info == null) {
             actions.addAll(listOf(
                 PleaseAction(project, label, "run",
-                    DefaultRunExecutor.getRunExecutorInstance(), newRunConfig(project, label)),
+                    DefaultRunExecutor.getRunExecutorInstance(), newRunConfig(project, target.pkg.pleaseRoot, label)),
                 PleaseAction(project, label, "run",
-                    DefaultDebugExecutor.getDebugExecutorInstance(), newRunConfig(project, label)),
+                    DefaultDebugExecutor.getDebugExecutorInstance(), newRunConfig(project, target.pkg.pleaseRoot, label)),
             ))
         }
 
         return actions
     }
 
-    private fun newRunConfig(project: Project, target: String) = PleaseRunConfiguration(
+    private fun newRunConfig(project: Project, pleaseRoot : String, target: String) = PleaseRunConfiguration(
         project,
         PleaseRunConfigurationType.Factory(PleaseRunConfigurationType()),
-        PleaseRunConfigArgs(target)
+        PleaseRunConfigArgs(target, pleaseRoot = pleaseRoot)
     )
 
-    private fun newBuildConfig(project: Project, target: String) = PleaseBuildConfiguration(
+    private fun newBuildConfig(project: Project, pleaseRoot : String, target: String) = PleaseBuildConfiguration(
         project,
         PleaseBuildConfigurationType.Factory(PleaseBuildConfigurationType()),
-        PleaseBuildConfigArgs(target)
+        PleaseBuildConfigArgs(target, pleaseRoot = pleaseRoot)
     )
 
-    fun newTestConfig(project: Project, target: String, tests: String = "") = PleaseTestConfiguration(
+    fun newTestConfig(project: Project, pleaseRoot : String, target: String, tests: String = "") = PleaseTestConfiguration(
         project,
         PleaseTestConfigurationType.Factory(PleaseTestConfigurationType()),
-        PleaseTestConfigArgs(target, tests=tests)
+        PleaseTestConfigArgs(target, tests = tests, pleaseRoot = pleaseRoot)
     )
 
 }
