+++ src/main/kotlin/net/thoughtmachine/please/plugin/runconfiguration/PleaseTestConfiguration.kt
@@ -18,6 +18,7 @@ import javax.swing.JPanel
 
 data class PleaseTestConfigArgs(
     var target: String,
+    var pleaseRoot: String = "",
     var pleaseArgs: String = "",
     var tests: String = ""
 )
@@ -40,6 +41,7 @@ class PleaseTestConfigurationType : ConfigurationTypeBase("PleaseTestConfigurati
 
 class PleaseTestConfigurationSettings : SettingsEditor<PleaseTestConfiguration>() {
     private val target = JBTextField("//some:target")
+    private val pleaseRoot = JBTextField()
     private val pleaseArgs = JBTextField()
     private val tests = JBTextField()
 
@@ -49,17 +51,20 @@ class PleaseTestConfigurationSettings : SettingsEditor<PleaseTestConfiguration>(
             .addLabeledComponent(JBLabel("Target: "), target, 1, false)
             .addLabeledComponent(JBLabel("Please args: "), pleaseArgs, 3, false)
             .addLabeledComponent(JBLabel("Tests: "), tests, 4, false)
+            .addLabeledComponent(JBLabel("Please project root: "), pleaseRoot, 4, false)
             .addComponentFillVertically(JPanel(), 0).panel
     }
 
     override fun applyEditorTo(s: PleaseTestConfiguration) {
         s.args.target = target.text
+        s.args.pleaseRoot = pleaseRoot.text
         s.args.pleaseArgs = pleaseArgs.text
         s.args.tests = tests.text
     }
 
     override fun resetEditorFrom(s: PleaseTestConfiguration) {
-        target.text = s.args.target.toString()
+        target.text = s.args.target
+        pleaseRoot.text = s.args.pleaseRoot
         pleaseArgs.text = s.args.pleaseArgs
         tests.text = s.args.tests
     }
@@ -79,22 +84,24 @@ class PleaseTestConfiguration(
 
     override fun target() = args.target
     override fun pleaseArgs() = args.pleaseArgs
+    override fun pleaseRoot() = args.pleaseRoot
 
     override fun getState(executor: Executor, environment: ExecutionEnvironment): RunProfileState {
         val plzArgs = Commandline.translateCommandline(args.pleaseArgs).toList()
         if (executor == PleaseBuildExecutor) {
-            return PleaseBuildConfiguration.getBuildProfileState(project, args.target, plzArgs)
+            return PleaseBuildConfiguration.getBuildProfileState(project, args.pleaseRoot, args.target, plzArgs)
         }
 
         if (executor == DefaultDebugExecutor.getDebugExecutorInstance()) {
             return PleaseDebugState(this, environment, computeDebugAddress(null))
         }
 
-        return PleaseProfileState(project, Please(project, pleaseArgs = plzArgs).test(args.target, args.tests))
+        return PleaseProfileState(project,  pleaseRoot = pleaseRoot(), Please(project, pleaseArgs = plzArgs).test(args.target, args.tests))
     }
 
     override fun writeExternal(element: Element) {
-        element.setAttribute("target", args.target.toString())
+        element.setAttribute("target", args.target)
+        element.setAttribute("pleaseRoot", args.pleaseRoot)
         element.setAttribute("pleaseArgs", args.pleaseArgs)
         element.setAttribute("tests", args.tests)
     }
@@ -102,6 +109,7 @@ class PleaseTestConfiguration(
     override fun readExternal(element: Element) {
         args = PleaseTestConfigArgs(
             target = element.getAttributeValue("target") ?: "//some:target",
+            pleaseRoot = element.getAttributeValue("pleaseRoot") ?: "",
             pleaseArgs = element.getAttributeValue("pleaseArgs") ?: "",
             tests = element.getAttributeValue("tests") ?: "",
         )
